+++ DSCResources/MSFT_xADComputer/MSFT_xADComputer.psm1
@@ -1,90 +1,125 @@
-$moduleRoot = Split-Path -Path $MyInvocation.MyCommand.Path -Parent
+$script:resourceModulePath = Split-Path -Path (Split-Path -Path $PSScriptRoot -Parent) -Parent
+$script:modulesFolderPath = Join-Path -Path $script:resourceModulePath -ChildPath 'Modules'
 
-# Import the common AD functions
-$adCommonFunctions = Join-Path `
-    -Path (Split-Path -Path $PSScriptRoot -Parent) `
-    -ChildPath '\MSFT_xADCommon\MSFT_xADCommon.psm1'
-Import-Module -Name $adCommonFunctions
+$script:localizationModulePath = Join-Path -Path $script:modulesFolderPath -ChildPath 'DscResource.LocalizationHelper'
+Import-Module -Name (Join-Path -Path $script:localizationModulePath -ChildPath 'DscResource.LocalizationHelper.psm1')
 
-#region LocalizedData
-$culture = 'en-us'
-if (Test-Path -Path (Join-Path -Path $moduleRoot -ChildPath $PSUICulture))
-{
-    $culture = $PSUICulture
-}
-$importLocalizedDataParams = @{
-    BindingVariable = 'LocalizedData'
-    Filename = 'MSFT_xADComputer.psd1'
-    BaseDirectory = $moduleRoot
-    UICulture = $culture
-}
-Import-LocalizedData @importLocalizedDataParams
-#endregion
+$script:dscResourcePath = Split-Path -Path $PSScriptRoot -Parent
+Import-Module -Name (Join-Path -Path $script:dscResourcePath -ChildPath '\MSFT_xADCommon\MSFT_xADCommon.psm1')
 
-# Create a property map that maps the DSC resource parameters to the
-# Active Directory computer attributes.
-$adPropertyMap = @(
-    @{ Parameter = 'ComputerName'; ADProperty = 'cn'; }
-    @{ Parameter = 'Location'; }
-    @{ Parameter = 'DnsHostName'; }
-    @{ Parameter = 'ServicePrincipalNames'; }
-    @{ Parameter = 'UserPrincipalName'; }
-    @{ Parameter = 'DisplayName'; }
-    @{ Parameter = 'Path'; ADProperty = 'distinguishedName'; }
-    @{ Parameter = 'Description'; }
-    @{ Parameter = 'Enabled'; }
-    @{ Parameter = 'Manager'; ADProperty = 'managedBy'; }
+$script:localizedData = Get-LocalizedData -ResourceName 'MSFT_xADComputer'
+
+<#
+    A property map that maps the resource parameters to the corresponding
+    Active Directory computer account object attribute.
+#>
+$script:computerObjectPropertyMap = @(
+    @{
+        ParameterName = 'ComputerName'
+        PropertyName  = 'CN'
+    },
+    @{
+        ParameterName = 'Location'
+    },
+    @{
+        ParameterName = 'DnsHostName'
+    },
+    @{
+        ParameterName = 'ServicePrincipalNames'
+    },
+    @{
+        ParameterName = 'UserPrincipalName'
+    },
+    @{
+        ParameterName = 'DisplayName'
+    },
+    @{
+        ParameterName = 'Path'
+        PropertyName  = 'DistinguishedName'
+    },
+    @{
+        ParameterName = 'Description'
+    },
+    @{
+        ParameterName = 'Enabled'
+    },
+    @{
+        ParameterName = 'Manager'
+        PropertyName  = 'ManagedBy'
+    },
+    @{
+        ParameterName = 'DistinguishedName'
+        ParameterType = 'Read'
+        PropertyName  = 'DistinguishedName'
+    },
+    @{
+        ParameterName = 'SID'
+        ParameterType = 'Read'
+    }
 )
 
+<#
+    .SYNOPSIS
+        Returns the current state of the Active Directory computer account.
 
+    .PARAMETER ComputerName
+         Specifies the name of the Active Directory computer account to manage.
+         You can identify a computer by its distinguished name, GUID, security
+         identifier (SID) or Security Accounts Manager (SAM) account name.
+
+    .PARAMETER RequestFile
+        Specifies the full path to the Offline Domain Join Request file to create.
+
+    .PARAMETER Enabled
+        DEPRECATED - DO NOT USE.
+
+        It is a parameter in Get-TargetResource to write the deprecated message.
+
+    .PARAMETER EnabledOnCreation
+        Specifies if the computer account is created enabled or disabled.
+        By default the computer account will be created using the default
+        value of the cmdlet New-ADComputer.
+
+    .PARAMETER DomainController
+        Specifies the Active Directory Domain Services instance to connect to perform the task.
+
+        Used by Get-ADCommonParameters and is returned as a common parameter.
+
+    .PARAMETER DomainAdministratorCredential
+        Specifies the user account credentials to use to perform the task.
+
+        Used by Get-ADCommonParameters and is returned as a common parameter.
+
+    .PARAMETER RestoreFromRecycleBin
+        Indicates whether or not the computer object should first tried to be
+        restored from the recycle bin before creating a new computer object.
+#>
 function Get-TargetResource
 {
     [CmdletBinding()]
     [OutputType([System.Collections.Hashtable])]
     param
     (
-        # Common Name
-        [Parameter(Mandatory)]
-        [System.String] $ComputerName,
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $ComputerName,
 
-        [ValidateSet('Present', 'Absent')]
-        [System.String] $Ensure = 'Present',
-
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $UserPrincipalName,
+        [System.String]
+        $RequestFile,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $DisplayName,
+        [System.Boolean]
+        $Enabled,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $Path,
+        [System.String]
+        $DomainController,
 
-        [ValidateNotNull()]
-        [System.String] $Location,
-
-        [ValidateNotNull()]
-        [System.String] $DnsHostName,
-
-        [ValidateNotNull()]
-        [System.String[]] $ServicePrincipalNames,
-
-        [ValidateNotNull()]
-        [System.String] $Description,
-
-        # Computer's manager specified as a Distinguished Name (DN)
-        [ValidateNotNull()]
-        [System.String] $Manager,
-
-        [ValidateNotNull()]
-        [System.String] $RequestFile,
-
-        [ValidateNotNull()]
-        [System.Boolean] $Enabled,
-
-        [ValidateNotNull()]
-        [System.String] $DomainController,
-
-        # Ideally this should just be called 'Credential' but is here for consistency with xADUser
+        [Parameter()]
         [ValidateNotNull()]
         [System.Management.Automation.PSCredential]
         [System.Management.Automation.CredentialAttribute()]
@@ -98,88 +133,97 @@ function Get-TargetResource
         [Parameter()]
         [ValidateNotNull()]
         [System.Boolean]
-        $CreateDisabled = $false
+        $EnabledOnCreation
     )
 
-    Assert-Module -ModuleName 'ActiveDirectory';
-    Import-Module -Name 'ActiveDirectory' -Verbose:$false;
+    <#
+        We have the deprecated message in Get-TargetResource so that it is
+        shown when both Test- and Set-TargetResource calls Get-TargetResource.
+    #>
+    if ($PSBoundParameters.ContainsKey('Enabled'))
+    {
+        Write-Warning -Message $script:localizedData.EnabledDeprecatedMessage
+    }
+
+    Assert-Module -ModuleName 'ActiveDirectory' -ImportModule
+
+    <#
+        These are properties that have no corresponding property in a
+        Computer account object.
+    #>
+    $getTargetResourceReturnValue = @{
+        Ensure                        = 'Absent'
+        ComputerName                  = $null
+        Location                      = $null
+        DnsHostName                   = $null
+        ServicePrincipalNames         = $null
+        UserPrincipalName             = $null
+        DisplayName                   = $null
+        Path                          = $null
+        Description                   = $null
+        Enabled                       = $false
+        Manager                       = $null
+        DomainController              = $DomainController
+        DomainAdministratorCredential = $DomainAdministratorCredential
+        RequestFile                   = $RequestFile
+        RestoreFromRecycleBin         = $RestoreFromRecycleBin
+        EnabledOnCreation             = $EnabledOnCreation
+        DistinguishedName             = $null
+        SID                           = $null
+        SamAccountName                = $null
+    }
+
+    $getADComputerResult = $null
 
     try
     {
-        $adCommonParameters = Get-ADCommonParameters @PSBoundParameters;
+        $commonParameters = Get-ADCommonParameters @PSBoundParameters
 
-        $adProperties = @();
-        # Create an array of the AD property names to retrieve from the property map
-        foreach ($property in $adPropertyMap)
-        {
+        <#
+            Create an array of the Active Directory Computer object property
+            names to retrieve from the Computer object.
+        #>
+        $computerObjectProperties = Convert-PropertyMapToObjectProperties -PropertyMap $script:computerObjectPropertyMap
 
-            if ($property.ADProperty)
-            {
-                $adProperties += $property.ADProperty;
-            }
-            else
-            {
-                $adProperties += $property.Parameter;
-            }
-        }
+        Write-Verbose -Message ($script:localizedData.RetrievingComputerAccount -f $ComputerName)
 
-        Write-Verbose -Message ($LocalizedData.RetrievingADComputer -f $ComputerName);
-        $adComputer = Get-ADComputer @adCommonParameters -Properties $adProperties;
-        Write-Verbose -Message ($LocalizedData.ADComputerIsPresent -f $ComputerName);
-        $Ensure = 'Present';
+        # If the computer account is not found Get-ADComputer will throw an error.
+        $getADComputerResult = Get-ADComputer @commonParameters -Properties $computerObjectProperties
+
+        Write-Verbose -Message ($script:localizedData.ComputerAccountIsPresent -f $ComputerName)
+
+        $getTargetResourceReturnValue['Ensure'] = 'Present'
+        $getTargetResourceReturnValue['ComputerName'] = $getADComputerResult.CN
+        $getTargetResourceReturnValue['Location'] = $getADComputerResult.Location
+        $getTargetResourceReturnValue['DnsHostName'] = $getADComputerResult.DnsHostName
+        $getTargetResourceReturnValue['ServicePrincipalNames'] = [System.String[]] $getADComputerResult.ServicePrincipalNames
+        $getTargetResourceReturnValue['UserPrincipalName'] = $getADComputerResult.UserPrincipalName
+        $getTargetResourceReturnValue['DisplayName'] = $getADComputerResult.DisplayName
+        $getTargetResourceReturnValue['Path'] = Get-ADObjectParentDN -DN $getADComputerResult.DistinguishedName
+        $getTargetResourceReturnValue['Description'] = $getADComputerResult.Description
+        $getTargetResourceReturnValue['Enabled'] = $getADComputerResult.Enabled
+        $getTargetResourceReturnValue['Manager'] = $getADComputerResult.ManagedBy
+        $getTargetResourceReturnValue['DomainController'] = $DomainController
+        $getTargetResourceReturnValue['DomainAdministratorCredential'] = $DomainAdministratorCredential
+        $getTargetResourceReturnValue['RequestFile'] = $RequestFile
+        $getTargetResourceReturnValue['RestoreFromRecycleBin'] = $RestoreFromRecycleBin
+        $getTargetResourceReturnValue['EnabledOnCreation'] = $EnabledOnCreation
+        $getTargetResourceReturnValue['DistinguishedName'] = $getADComputerResult.DistinguishedName
+        $getTargetResourceReturnValue['SID'] = $getADComputerResult.SID
+        $getTargetResourceReturnValue['SamAccountName'] = $getADComputerResult.SamAccountName
     }
     catch [Microsoft.ActiveDirectory.Management.ADIdentityNotFoundException]
     {
-        Write-Verbose -Message ($LocalizedData.ADComputerNotPresent -f $ComputerName);
-        $Ensure = 'Absent';
+        Write-Verbose -Message ($script:localizedData.ComputerAccountIsAbsent -f $ComputerName)
     }
     catch
     {
-        Write-Error -Message ($LocalizedData.RetrievingADComputerError -f $ComputerName);
-        throw $_;
+        $errorMessage = $script:localizedData.FailedToRetrieveComputerAccount -f $ComputerName
+        New-InvalidOperationException -Message $errorMessage -ErrorRecord $_
     }
 
-    $targetResource = @{
-        ComputerName      = $ComputerName;
-        DistinguishedName = $adComputer.DistinguishedName; # Read-only property
-        SID               = $adComputer.SID; # Read-only property
-        Ensure            = $Ensure;
-        DomainController  = $DomainController;
-        RequestFile    = $RequestFile;
-    }
-
-    # Retrieve each property from the ADPropertyMap and add to the hashtable
-    foreach ($property in $adPropertyMap)
-    {
-        $propertyName = $property.Parameter;
-        if ($propertyName -eq 'Path') {
-            # The path returned is not the parent container
-            if (-not [System.String]::IsNullOrEmpty($adComputer.DistinguishedName))
-            {
-                $targetResource['Path'] = Get-ADObjectParentDN -DN $adComputer.DistinguishedName;
-            }
-        }
-        elseif ($property.ADProperty)
-        {
-            # The AD property name is different to the function parameter to use this
-            $targetResource[$propertyName] = $adComputer.($property.ADProperty);
-        }
-        else
-        {
-            # The AD property name matches the function parameter
-            if ($adComputer.$propertyName -is [Microsoft.ActiveDirectory.Management.ADPropertyValueCollection])
-            {
-                $targetResource[$propertyName] = $adComputer.$propertyName -as [System.String[]];
-            }
-            else
-            {
-                $targetResource[$propertyName] = $adComputer.$propertyName;
-            }
-        }
-    }
-    return $targetResource;
-
-} #end function Get-TargetResource
+    return $getTargetResourceReturnValue
+}
 
 
 function Test-TargetResource
@@ -189,52 +233,79 @@ function Test-TargetResource
     param
     (
         # Common Name
-        [Parameter(Mandatory)]
-        [System.String] $ComputerName,
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $ComputerName,
 
+        [Parameter()]
         [ValidateSet('Present', 'Absent')]
-        [System.String] $Ensure = 'Present',
+        [System.String]
+        $Ensure = 'Present',
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $UserPrincipalName,
+        [System.String]
+        $UserPrincipalName,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $DisplayName,
+        [System.String]
+        $DisplayName,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $Path,
+        [System.String]
+        $Path,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $Location,
+        [System.String]
+        $Location,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $DnsHostName,
+        [System.String]
+        $DnsHostName,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String[]] $ServicePrincipalNames,
+        [System.String[]]
+        $ServicePrincipalNames,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $Description,
+        [System.String]
+        $Description,
 
         # Computer's manager specified as a Distinguished Name (DN)
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $Manager,
+        [System.String]
+        $Manager,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $RequestFile,
+        [System.String]
+        $RequestFile,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.Boolean] $Enabled,
+        [System.Boolean]
+        $Enabled,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $DomainController,
+        [System.String]
+        $DomainController,
 
         # Ideally this should just be called 'Credential' but is here for backwards compatibility
+        [Parameter()]
         [ValidateNotNull()]
         [System.Management.Automation.PSCredential]
         [System.Management.Automation.CredentialAttribute()]
         $DomainAdministratorCredential,
 
+        [Parameter()]
         [ValidateNotNull()]
         [System.Boolean]
         $RestoreFromRecycleBin,
@@ -242,25 +313,25 @@ function Test-TargetResource
         [Parameter()]
         [ValidateNotNull()]
         [System.Boolean]
-        $CreateDisabled = $false
+        $EnabledOnCreation
     )
 
-    $targetResource = Get-TargetResource @PSBoundParameters;
-    $isCompliant = $true;
+    $targetResource = Get-TargetResource @PSBoundParameters
+    $isCompliant = $true
 
     if ($Ensure -eq 'Absent')
     {
         if ($targetResource.Ensure -eq 'Present')
         {
-            Write-Verbose -Message ($LocalizedData.ADComputerNotDesiredPropertyState -f `
-                                    'Ensure', $PSBoundParameters.Ensure, $targetResource.Ensure);
-            $isCompliant = $false;
+            Write-Verbose -Message ($script:localizedData.ADComputerNotDesiredPropertyState -f `
+                    'Ensure', $PSBoundParameters.Ensure, $targetResource.Ensure)
+            $isCompliant = $false
         }
     }
     else
     {
         # Add ensure as it may not be explicitly passed and we want to enumerate it.
-        $PSBoundParameters['Ensure'] = $Ensure;
+        $PSBoundParameters['Ensure'] = $Ensure
 
         foreach ($parameter in $PSBoundParameters.Keys)
         {
@@ -275,23 +346,23 @@ function Test-TargetResource
                 elseif ($parameter -eq 'ServicePrincipalNames')
                 {
                     $testMembersParams = @{
-                        ExistingMembers = $targetResource.ServicePrincipalNames -as [System.String[]];
-                        Members = $ServicePrincipalNames;
+                        ExistingMembers = $targetResource.ServicePrincipalNames -as [System.String[]]
+                        Members         = $ServicePrincipalNames
                     }
                     if (-not (Test-Members @testMembersParams))
                     {
-                        $existingSPNs = $testMembersParams['ExistingMembers'] -join ',';
-                        $desiredSPNs = $ServicePrincipalNames -join ',';
-                        Write-Verbose -Message ($LocalizedData.ADComputerNotDesiredPropertyState -f `
-                                                'ServicePrincipalNames', $desiredSPNs, $existingSPNs);
-                        $isCompliant = $false;
+                        $existingSPNs = $testMembersParams['ExistingMembers'] -join ','
+                        $desiredSPNs = $ServicePrincipalNames -join ','
+                        Write-Verbose -Message ($script:localizedData.ADComputerNotDesiredPropertyState -f `
+                                'ServicePrincipalNames', $desiredSPNs, $existingSPNs)
+                        $isCompliant = $false
                     }
                 }
                 elseif ($PSBoundParameters.$parameter -ne $targetResource.$parameter)
                 {
-                    Write-Verbose -Message ($LocalizedData.ADComputerNotDesiredPropertyState -f `
-                                            $parameter, $PSBoundParameters.$parameter, $targetResource.$parameter);
-                    $isCompliant = $false;
+                    Write-Verbose -Message ($script:localizedData.ADComputerNotDesiredPropertyState -f `
+                            $parameter, $PSBoundParameters.$parameter, $targetResource.$parameter)
+                    $isCompliant = $false
                 }
             }
         } #end foreach PSBoundParameter
@@ -299,12 +370,12 @@ function Test-TargetResource
 
     if ($isCompliant)
     {
-        Write-Verbose -Message ($LocalizedData.ADComputerInDesiredState -f $ComputerName)
+        Write-Verbose -Message ($script:localizedData.ADComputerInDesiredState -f $ComputerName)
         return $true
     }
     else
     {
-        Write-Verbose -Message ($LocalizedData.ADComputerNotInDesiredState -f $ComputerName)
+        Write-Verbose -Message ($script:localizedData.ADComputerNotInDesiredState -f $ComputerName)
         return $false
     }
 
@@ -317,52 +388,79 @@ function Set-TargetResource
     param
     (
         # Common Name
-        [Parameter(Mandatory)]
-        [System.String] $ComputerName,
+        [Parameter(Mandatory = $true)]
+        [System.String]
+        $ComputerName,
 
+        [Parameter()]
         [ValidateSet('Present', 'Absent')]
-        [System.String] $Ensure = 'Present',
+        [System.String]
+        $Ensure = 'Present',
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $UserPrincipalName,
+        [System.String]
+        $UserPrincipalName,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $DisplayName,
+        [System.String]
+        $DisplayName,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $Path,
+        [System.String]
+        $Path,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $Location,
+        [System.String]
+        $Location,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $DnsHostName,
+        [System.String]
+        $DnsHostName,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String[]] $ServicePrincipalNames,
+        [System.String[]]
+        $ServicePrincipalNames,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $Description,
+        [System.String]
+        $Description,
 
         # Computer's manager specified as a Distinguished Name (DN)
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $Manager,
+        [System.String]
+        $Manager,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $RequestFile,
+        [System.String]
+        $RequestFile,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.Boolean] $Enabled,
+        [System.Boolean]
+        $Enabled,
 
+        [Parameter()]
         [ValidateNotNull()]
-        [System.String] $DomainController,
+        [System.String]
+        $DomainController,
 
         # Ideally this should just be called 'Credential' but is here for backwards compatibility
+        [Parameter()]
         [ValidateNotNull()]
         [System.Management.Automation.PSCredential]
         [System.Management.Automation.CredentialAttribute()]
         $DomainAdministratorCredential,
 
+        [Parameter()]
         [ValidateNotNull()]
         [System.Boolean]
         $RestoreFromRecycleBin,
@@ -370,21 +468,22 @@ function Set-TargetResource
         [Parameter()]
         [ValidateNotNull()]
         [System.Boolean]
-        $CreateDisabled = $false
+        $EnabledOnCreation
     )
 
-    $targetResource = Get-TargetResource @PSBoundParameters;
+    $targetResource = Get-TargetResource @PSBoundParameters
 
     ## Add ensure as they may not be explicitly passed and we want to enumerate them
-    $PSBoundParameters['Ensure'] = $Ensure;
+    $PSBoundParameters['Ensure'] = $Ensure
 
     if ($Ensure -eq 'Present')
     {
-        if ($targetResource.Ensure -eq 'Absent') {
+        if ($targetResource.Ensure -eq 'Absent')
+        {
             # Try to restore account if it exists
-            if($RestoreFromRecycleBin)
+            if ($RestoreFromRecycleBin)
             {
-                Write-Verbose -Message ($LocalizedData.RestoringADComputer -f $ComputerName)
+                Write-Verbose -Message ($script:localizedData.RestoringADComputer -f $ComputerName)
                 $restoreParams = Get-ADCommonParameters @PSBoundParameters
                 $restorationSuccessful = Restore-ADCommonObject @restoreParams -ObjectClass Computer -ErrorAction Stop
             }
@@ -398,60 +497,60 @@ function Set-TargetResource
                 if ($RequestFile)
                 {
                     # Use DJOIN to create the computer account as well as the ODJ Request file.
-                    Write-Verbose -Message ($LocalizedData.ODJRequestStartMessage -f `
-                            $DomainName,$ComputerName,$RequestFile)
+                    Write-Verbose -Message ($script:localizedData.ODJRequestStartMessage -f `
+                            $DomainName, $ComputerName, $RequestFile)
 
                     # This should only be performed on a Domain Member, so detect the Domain Name.
                     $DomainName = Get-DomainName
                     $DJoinParameters = @(
                         '/PROVISION'
-                        '/DOMAIN',$DomainName
-                        '/MACHINE',$ComputerName )
+                        '/DOMAIN', $DomainName
+                        '/MACHINE', $ComputerName )
                     if ($PSBoundParameters.ContainsKey('Path'))
                     {
-                        $DJoinParameters += @( '/MACHINEOU',$Path )
+                        $DJoinParameters += @( '/MACHINEOU', $Path )
                     } # if
 
                     if ($PSBoundParameters.ContainsKey('DomainController'))
                     {
-                        $DJoinParameters += @( '/DCNAME',$DomainController )
+                        $DJoinParameters += @( '/DCNAME', $DomainController )
                     } # if
 
-                    $DJoinParameters += @( '/SAVEFILE',$RequestFile )
+                    $DJoinParameters += @( '/SAVEFILE', $RequestFile )
                     $Result = & djoin.exe @DjoinParameters
 
                     if ($LASTEXITCODE -ne 0)
                     {
                         $errorId = 'ODJRequestError'
-                        $errorMessage = $($LocalizedData.ODJRequestError `
-                            -f $LASTEXITCODE,$Result)
+                        $errorMessage = $($script:localizedData.ODJRequestError `
+                                -f $LASTEXITCODE, $Result)
                         ThrowInvalidOperationError -ErrorId $errorId -ErrorMessage $errorMessage
                     } # if
 
-                    Write-Verbose -Message ($LocalizedData.ODJRequestCompleteMessage -f `
-                            $DomainName,$ComputerName,$RequestFile)
+                    Write-Verbose -Message ($script:localizedData.ODJRequestCompleteMessage -f `
+                            $DomainName, $ComputerName, $RequestFile)
                 }
                 else
                 {
                     # Create the computer account using New-ADComputer
-                    $newADComputerParams = Get-ADCommonParameters @PSBoundParameters -UseNameParameter;
+                    $newADComputerParams = Get-ADCommonParameters @PSBoundParameters -UseNameParameter
                     if ($PSBoundParameters.ContainsKey('Path'))
                     {
-                        Write-Verbose -Message ($LocalizedData.UpdatingADComputerProperty -f 'Path', $Path);
-                        $newADComputerParams['Path'] = $Path;
+                        Write-Verbose -Message ($script:localizedData.UpdatingADComputerProperty -f 'Path', $Path)
+                        $newADComputerParams['Path'] = $Path
                     }
-                    Write-Verbose -Message ($LocalizedData.AddingADComputer -f $ComputerName);
-                    New-ADComputer @newADComputerParams;
+                    Write-Verbose -Message ($script:localizedData.AddingADComputer -f $ComputerName)
+                    New-ADComputer @newADComputerParams
                 } # if
             }
             else
             {
                 ## Create the computer account using New-ADComputer
-                $newADComputerParams = Get-ADCommonParameters @PSBoundParameters -UseNameParameter;
+                $newADComputerParams = Get-ADCommonParameters @PSBoundParameters -UseNameParameter
                 if ($PSBoundParameters.ContainsKey('Path'))
                 {
-                    Write-Verbose -Message ($LocalizedData.UpdatingADComputerProperty -f 'Path', $Path);
-                    $newADComputerParams['Path'] = $Path;
+                    Write-Verbose -Message ($script:localizedData.UpdatingADComputerProperty -f 'Path', $Path)
+                    $newADComputerParams['Path'] = $Path
                 }
 
                 <#
@@ -461,27 +560,27 @@ function Set-TargetResource
                 #>
                 if (
                     ($CreateDisabled -and -not $PSBoundParameters.ContainsKey('Enabled')) `
-                    -or ($PSBoundParameters.ContainsKey('Enabled') -and -not $Enabled)
+                        -or ($PSBoundParameters.ContainsKey('Enabled') -and -not $Enabled)
                 )
                 {
-                    Write-Verbose -Message ($LocalizedData.AddingADComputerAsDisabled -f $ComputerName);
-                    $newADComputerParams['Enabled'] = $false;
+                    Write-Verbose -Message ($script:localizedData.AddingADComputerAsDisabled -f $ComputerName)
+                    $newADComputerParams['Enabled'] = $false
                 }
                 else
                 {
-                    Write-Verbose -Message ($LocalizedData.AddingADComputer -f $ComputerName);
-                    $newADComputerParams['Enabled'] = $true;
+                    Write-Verbose -Message ($script:localizedData.AddingADComputer -f $ComputerName)
+                    $newADComputerParams['Enabled'] = $true
                 }
 
-                New-ADComputer @newADComputerParams;
+                New-ADComputer @newADComputerParams
             } # if
             ## Now retrieve the newly created computer
-            $targetResource = Get-TargetResource @PSBoundParameters;
+            $targetResource = Get-TargetResource @PSBoundParameters
         }
 
-        $setADComputerParams = Get-ADCommonParameters @PSBoundParameters;
-        $replaceComputerProperties = @{};
-        $removeComputerProperties = @{};
+        $setADComputerParams = Get-ADCommonParameters @PSBoundParameters
+        $replaceComputerProperties = @{ }
+        $removeComputerProperties = @{ }
         foreach ($parameter in $PSBoundParameters.Keys)
         {
             # Only check/action properties specified/declared parameters that match one of the function's
@@ -491,23 +590,23 @@ function Set-TargetResource
                 if ($parameter -eq 'Path' -and ($PSBoundParameters.Path -ne $targetResource.Path))
                 {
                     # Cannot move computers by updating the DistinguishedName property
-                    $adCommonParameters = Get-ADCommonParameters @PSBoundParameters;
+                    $commonParameters = Get-ADCommonParameters @PSBoundParameters
                     # Using the SamAccountName for identity with Move-ADObject does not work, use the DN instead
-                    $adCommonParameters['Identity'] = $targetResource.DistinguishedName;
-                    Write-Verbose -Message ($LocalizedData.MovingADComputer -f `
-                                            $targetResource.Path, $PSBoundParameters.Path);
-                    Move-ADObject @adCommonParameters -TargetPath $PSBoundParameters.Path;
+                    $commonParameters['Identity'] = $targetResource.DistinguishedName
+                    Write-Verbose -Message ($script:localizedData.MovingADComputer -f `
+                            $targetResource.Path, $PSBoundParameters.Path)
+                    Move-ADObject @commonParameters -TargetPath $PSBoundParameters.Path
                 }
                 elseif ($parameter -eq 'ServicePrincipalNames')
                 {
-                    Write-Verbose -Message ($LocalizedData.UpdatingADComputerProperty -f `
-                                            'ServicePrincipalNames', ($ServicePrincipalNames -join ','));
-                    $replaceComputerProperties['ServicePrincipalName'] = $ServicePrincipalNames;
+                    Write-Verbose -Message ($script:localizedData.UpdatingADComputerProperty -f `
+                            'ServicePrincipalNames', ($ServicePrincipalNames -join ','))
+                    $replaceComputerProperties['ServicePrincipalName'] = $ServicePrincipalNames
                 }
                 elseif ($parameter -eq 'Enabled' -and ($PSBoundParameters.$parameter -ne $targetResource.$parameter))
                 {
-                    Write-Verbose -Message ($LocalizedData.UpdatingADComputerProperty -f `
-                                            $parameter, $PSBoundParameters.$parameter);
+                    Write-Verbose -Message ($script:localizedData.UpdatingADComputerProperty -f `
+                            $parameter, $PSBoundParameters.$parameter)
 
                     <#
                         The Enabled property cannot be set as a hash table value in the
@@ -519,7 +618,7 @@ function Set-TargetResource
                 elseif ($PSBoundParameters.$parameter -ne $targetResource.$parameter)
                 {
                     # Find the associated AD property
-                    $adProperty = $adPropertyMap | Where-Object { $_.Parameter -eq $parameter };
+                    $adProperty = $script:computerObjectPropertyMap | Where-Object { $_.Parameter -eq $parameter }
 
                     if ([System.String]::IsNullOrEmpty($adProperty))
                     {
@@ -531,40 +630,40 @@ function Set-TargetResource
                         # Only remove if the existing value in not null or empty
                         if (-not ([System.String]::IsNullOrEmpty($targetResource.$parameter)))
                         {
-                            Write-Verbose -Message ($LocalizedData.RemovingADComputerProperty -f `
-                                                    $parameter, $PSBoundParameters.$parameter);
+                            Write-Verbose -Message ($script:localizedData.RemovingADComputerProperty -f `
+                                    $parameter, $PSBoundParameters.$parameter)
                             if ($adProperty.UseCmdletParameter -eq $true)
                             {
                                 # We need to pass the parameter explicitly to Set-ADComputer, not via -Remove
-                                $setADComputerParams[$adProperty.Parameter] = $PSBoundParameters.$parameter;
+                                $setADComputerParams[$adProperty.Parameter] = $PSBoundParameters.$parameter
                             }
                             elseif ([System.String]::IsNullOrEmpty($adProperty.ADProperty))
                             {
-                                $removeComputerProperties[$adProperty.Parameter] = $targetResource.$parameter;
+                                $removeComputerProperties[$adProperty.Parameter] = $targetResource.$parameter
                             }
                             else
                             {
-                                $removeComputerProperties[$adProperty.ADProperty] = $targetResource.$parameter;
+                                $removeComputerProperties[$adProperty.ADProperty] = $targetResource.$parameter
                             }
                         }
                     } #end if remove existing value
                     else
                     {
                         # We are replacing the existing value
-                        Write-Verbose -Message ($LocalizedData.UpdatingADComputerProperty -f `
-                                                $parameter, $PSBoundParameters.$parameter);
+                        Write-Verbose -Message ($script:localizedData.UpdatingADComputerProperty -f `
+                                $parameter, $PSBoundParameters.$parameter)
                         if ($adProperty.UseCmdletParameter -eq $true)
                         {
                             # We need to pass the parameter explicitly to Set-ADComputer, not via -Replace
-                            $setADComputerParams[$adProperty.Parameter] = $PSBoundParameters.$parameter;
+                            $setADComputerParams[$adProperty.Parameter] = $PSBoundParameters.$parameter
                         }
                         elseif ([System.String]::IsNullOrEmpty($adProperty.ADProperty))
                         {
-                            $replaceComputerProperties[$adProperty.Parameter] = $PSBoundParameters.$parameter;
+                            $replaceComputerProperties[$adProperty.Parameter] = $PSBoundParameters.$parameter
                         }
                         else
                         {
-                            $replaceComputerProperties[$adProperty.ADProperty] = $PSBoundParameters.$parameter;
+                            $replaceComputerProperties[$adProperty.ADProperty] = $PSBoundParameters.$parameter
                         }
                     } #end if replace existing value
                 }
@@ -582,23 +681,23 @@ function Set-TargetResource
             ## Only pass -Remove and/or -Replace if we have something to set/change
             if ($replaceComputerProperties.Count -gt 0)
             {
-                $setADComputerParams['Replace'] = $replaceComputerProperties;
+                $setADComputerParams['Replace'] = $replaceComputerProperties
             }
             if ($removeComputerProperties.Count -gt 0)
             {
-                $setADComputerParams['Remove'] = $removeComputerProperties;
+                $setADComputerParams['Remove'] = $removeComputerProperties
             }
 
-            Write-Verbose -Message ($LocalizedData.UpdatingADComputer -f $ComputerName);
+            Write-Verbose -Message ($script:localizedData.UpdatingADComputer -f $ComputerName)
             Set-DscADComputer -SetADComputerParameters $setADComputerParams
         }
     }
     elseif (($Ensure -eq 'Absent') -and ($targetResource.Ensure -eq 'Present'))
     {
         # User exists and needs removing
-        Write-Verbose ($LocalizedData.RemovingADComputer -f $ComputerName);
-        $adCommonParameters = Get-ADCommonParameters @PSBoundParameters;
-        [ref] $null = Remove-ADComputer @adCommonParameters -Confirm:$false;
+        Write-Verbose ($script:localizedData.RemovingADComputer -f $ComputerName)
+        $commonParameters = Get-ADCommonParameters @PSBoundParameters
+        [ref] $null = Remove-ADComputer @commonParameters -Confirm:$false
     }
 
 } #end function Set-TargetResource